requirements.txt
ccxt==3.1.17
loguru==0.7.2
starknet-py==0.18.2
tabulate==0.9.0
requests==2.31.0
tqdm==4.66.1
web3==6.11.1
python-dotenv==1.0.0
termcolor==2.3.0
